requirements.txt
decord==0.6.0
diffusers==0.27.0
einops==0.8.0
imageio==2.34.0
matplotlib==3.8.3
numpy==1.24.4
omegaconf==2.3.0
onnxruntime_gpu==1.18.0
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
Pillow==11.0.0
pytlsd==0.0.2
safetensors==0.4.5
scikit_learn==1.6.0
seaborn==0.13.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.2
transformers==4.37.2
